requirements.txt
torch==1.9.0
gradio==2.3.7
faiss-cpu==1.7.1
sentence-transformers==2.0.0
langchain==0.0.1